Renaming translation_key to $1_fragment
Instead of `translation_key` which implies a full key, I want to
indicate that I am working with a translation key fragment. This also
aligns with the idea that there are different translation contexts to
consider for a predicate (i.e. translation for input hint form,
translation for input label, translation for display label).