Working on Translations for Predicates
Given that a predicate can span multiple concerns (i.e. the "role"
predicate as it applies to a book, movie, game, dataset, etc.). For
each of those concerns there is a nuance to the meaning. The
Milan::TranslationAssistant.for_predicate attempts to encapsulate the
expected structure and provide insight into how that structure
resolves.